+++ R/Index_calculations.r
@@ -479,11 +479,8 @@ poppr <- function(dat, total = TRUE, sublist = "ALL", blacklist = NULL,
 #' # Obtain a list of fstat files from a directory.
 #' x <- getfile(multi=TRUE, pattern="^.+?dat$")
 #'
-#' # set the working directory to that directory.
-#' setwd(x$path)
-#'
 #' # run the analysis on each file.
-#' poppr.all(x$files)
+#' poppr.all(file.path(x$path, x$files))
 #' }
 #==============================================================================# 
 poppr.all <- function(filelist, ...){