+++ R/internal_methods.R
@@ -220,7 +220,8 @@ mll.levels.internal <- function(x, set = TRUE, value){
 #==============================================================================#
 mlg.filter.internal <- function(gid, threshold = 0.0, missing = "asis", 
                                 memory = FALSE, algorithm = "farthest_neighbor", 
-                                distance = "diss.dist", threads = 1L, 
+                                distance = "diss.dist", denv = .GlobalEnv,
+                                threads = 1L, 
                                 stats = "MLGs", the_call = match.call(), ...){
 
   if (threads != 1L){
@@ -229,17 +230,20 @@ mlg.filter.internal <- function(gid, threshold = 0.0, missing = "asis",
   }
   # This will return a vector indicating the multilocus genotypes after applying
   # a minimum required distance threshold between multilocus genotypes.
-  if (is.character(distance) || is.function(distance)) {
+  dist_is_fun <- is.function(distance)
+  if (is.character(distance) || dist_is_fun) {
     if (memory==TRUE && identical(c(gid, distance, ...), .last.value.param$get())){
       dis <- .last.value.dist$get()
     } else {
       if (is.genind(gid)) {
-        the_dist <- as.character(the_call[["distance"]])
+        the_dist <- if (!dist_is_fun) as.character(the_call[["distance"]])
         call_len <- length(the_dist)
         is_diss_dist <- the_dist %in% "diss.dist"
-        any_dist <- the_dist %in% c("diss.dist", "nei.dist", "prevosti.dist",
-                                    "edwards.dist", "reynolds.dist", 
-                                    "rogers.dist", "provesti.dist")
+
+        dists <- c("diss.dist", "nei.dist", "prevosti.dist", "edwards.dist",
+                   "reynolds.dist", "rogers.dist", "provesti.dist")
+        any_dist <- the_dist %in% dists
+
         if (missing == "mean" && call_len == 1 && is_diss_dist){
           disswarn <- paste("Cannot use function diss.dist and correct for", 
                             "mean values.", "diss.dist will automatically",
@@ -255,7 +259,8 @@ mlg.filter.internal <- function(gid, threshold = 0.0, missing = "asis",
       } else {
         mpop <- gid
       }
-      DISTFUN <- match.fun(distance)
+      # browser()
+      DISTFUN <- if (!is.function(distance)) get(distance, envir = denv) else distance
       dis <- DISTFUN(mpop, ...)
       dis <- as.matrix(dis)
       if (memory == TRUE)
@@ -322,6 +327,7 @@ mlg.filter.internal <- function(gid, threshold = 0.0, missing = "asis",
     stop("Threshold must be a numeric or integer value", .call = FALSE)
   } 
     # Stats must be logical
+  # browser()
   STATARGS <- c("MLGS", "THRESHOLDS", "DISTANCES", "SIZES", "ALL")
   stats <- match.arg(toupper(stats), STATARGS, several.ok = TRUE)